a game for guessing numbers
You specify a number in your mind and you do not tell the computer this number. The program guesses a number and prints it. If it is greater than the number you have in your mind, you tell the program by typing the letter k that the number in your mind is less than the number the program printed and the program should guess and display another number. Or if the printed number is less than the number you have in your mind, you tell the program by typing the letter b that the number in your mind is greater than the number the program printed and the program should guess and display another number. Or if the printed number is the same number as the number in your mind, you tell the program by typing the letter d that it guessed correctly and the program ends.